Nexus audit iteration 1: fix slot indexing, combat, spell, stamina, and VM bugs
Seven gameplay bugs fixed using DM.BIN disassembly as reference:
- mechanics: 3 slot off-by-one (defense, ring, weapon used slots[n-1] not slots[n])
- combat: remove backwards wound penalty on creature attacks
- spell_effects: DISPEL removes all statuses, not only poison
- champions: linear stamina scaling (DM.BIN 0x029F38)
- click_route: slot clearing sign mismatch (-1 vs 255)
- script_vm: uint32_t return for SH-2 high-RAM addresses
- spawner: timer init to respawn_delay, not 0

+++ src/nexus/nexus_v1_click_route.c
@@ -92,7 +92,7 @@ static Nexus_ClickResult click_route_inventory(Nexus_MechanicsState *st,
         for (i = 0; i < NEXUS_INVENTORY_SLOTS; i++) {
             if (champ->inventory[i] == 0xFFU) {
                 champ->inventory[i] = (uint8_t)item_id;
-                champ->slots[equip_slot] = (uint8_t)-1;
+                champ->slots[equip_slot] = -1;
                 nexus_champion_recalc_load(champ);
                 return NEXUS_CLICK_RESULT_OK;
             }
